Add tracking of memory_pgpgin and memory_pgpgout
Per [the docs][1]:

**`memory_pgpgout` (Pages Written to Disk):**

The cumulative total of the pages written to disk. Sudden high
variations on this number can indicate short duration spikes in swap
usage. The other memory related metrics are point in time snapshots and
can miss short spikes.

**`memory_pgpgin` (Pages Read from Disk):**

The cumulative total of the pages read from disk. As with the previous
metric, watch out for sudden variations.

[1]: https://devcenter.heroku.com/articles/log-runtime-metrics#memory-swap